Replace supertab with a mapping
I found supertab to be too intrusive for my day-to-day workflow. It
turns out that I expect ctrl+space to open up the autocompletion menu,
not tab. So, I'm removing supertab and replacing it with above mapping.

I would also like to map tab to insert the currently selected
autocompletion, but that will wait for another day.