Split up .vimrc
In the interest of cleaning up my .vimrc, I want to organize this a bit
better. Vim allows for the organization of filetype-specific
configuration in the ftplugin directory. I may as well take advantage of
this feature.

I am also abusing the plugin directory as a place to stick my settings,
since it is autoloaded and pathogen takes care of my actual vim plugins.

Some adjustments were made to the configuration at the same time,
inspired by Wincent's configuration:

  https://github.com/wincent/wincent