Add visual block reselection after in-/out-dent
Although it is easy enough to repeat the previous action in vim by using
`.`, I find myself often wanting to do something different on the same
visual block after indenting it or outdenting it. This commit adds a
visual mode mapping that reselects the visual block after indenting or
outdenting.

Taken from:

  http://vimbits.com/bits/20