Automatically disable paste mode when leaving insert
When pasting content, I usually only want to paste once, and then turn
off paste mode. The process typically looks like this:

  :set paste
  i<C-v><ESC>
  :set nopaste

This commit adds an autocmd that automates the last step of this. This
may end up being annoying in the instance where I want to paste multiple
things, but I am having a hard time recalling when I have actually
wanted to do that at this time so I'm not going to worry about it.