Vim: remap Q to @q
Pressing `Q` in normal mode by default enters ex mode, which is entirely
useless to me and usually ends up happening accidentally and wasting
time. This commit remaps `Q` to `@q`, which plays back the macro stored
in `q`. This allows me to quickly record a macro with `qq` and then play
it back again with `Q`.

Inspiration from:

  http://www.reddit.com/r/vim/comments/1v2agp/ex_mode_should_i_learn_it_or_just_remap_shiftq_so/