Vim: add vim-multiple-cursors plugin
This plugin implements Sublime Text's multiple cursors functionality for
vim, which is something that I have been missing quite a bit since
switching. After taking it for a spin, it seems to work quite well.

  Ctrl+n: select current word, or next match
  Ctrl+p: un-select current match and go back
  Ctrl+x: un-select current match and go next

There appear to be some options available that I may end up wanting to
use:

  g:multi_cursor_exit_from_visual_mode = 0
  g:multi_cursor_exit_from_insert_mode = 0

But for now I'm going to roll with the default settings and see how it
goes.