Vim: set nocompatible
According to a post on stackoverflow[1], `set nocompatible` isn't
completely useless. Adding it back to my .vimrc to be safe, rather than
sorry.

Note that this must come before other `set` calls because setting this
magically wipes out anything that was set before it. Therefore, I put
this as the first line in my .vimrc instead of where all of my other
settings are, which is currently .vim/plugin/settings.vim.

[1]: http://stackoverflow.com/a/5845583/18986