Make italics work in iTerm2, tmux, and vim
The .terminfo files and install.sh script were taken from
https://github.com/wincent/wincent/tree/master/extras.

I needed to bring in a version of Inconsolata that included italics to
make this work.

To bring this all together, I needed to use the italics branch of
@wincent's fork of the vim-colors-solarized repo, but I wanted it to
work well with my update script which expects all of my submodules to be
on master so I forked it and cherry-picked the needed commit onto
master.