Vim: Simplify create class snippet
We no longer add use strict at the top of our JS files because it is
added automatically by our ES6 transpiler. Additionally, I found it
difficult to remember to use this snippet, and think that it might be
easier to remember if it were scoped to the function that I would be
calling, which is React.createClass.