Vim: Add snippet for JS function def with dynamic docblock
This is an experiment. I have been creating a lot of JavaScript
functions and then going back to add @param tags to document arguments
and their data types. I'm wondering if I can streamline this by using a
snippet.

This snippet will parse the arguments list and automatically create
@param tags for me in a JSDoc docblock. I made it smart enough to not
create a docblock if there are no arguments since it is easy enough to
generate one with docblock specific snippets if I want it for a
description or a @return tag or something.

I had wanted to make dynamic tabstops for the datatype portion of the
@param tags, but Ultisnips does not currently support this feature, so
I'll have to settle for manually going back and adding these.

Also, since this is an experiment for me, I decided to use the function
definition style that I use the most: functions that are properties of
objects.

- https://answers.launchpad.net/ultisnips/+question/187774
- https://answers.launchpad.net/ultisnips/+question/215812
- https://github.com/SirVer/ultisnips/issues/235