Vim: use PascalCased filename in jsx createClass snippet
We name our JSX files like component_name.js.jsx, and our components
themselves like ComponentName. This commit adds some intelligence to my
createClass JSX snippet to automatically convert the filename from
snake_case to PascalCase for the module definition and export.

This line is pretty long and there is a lot going on in there, but I
think it is fine for this snippet for now. If it grows much larger, I
may want to find a way to break it up a bit.