Vim: prevent supertab from eating <CR>
I had added supertab in a supporting role to help make my YouCompleteMe
and UltiSnips integration work better. However, as I was using it, I
noticed that sometimes when I pressed enter, vim would only close the
completion popup and not add a newline to the buffer. This was a pretty
jarring experience for me and ended up regularly slowing me down.

I was able to trace this back a feature in supertab that is enabled by
default. Disabling `g:SuperTabCrMapping` makes vim behave more like I
expect it.

> When enabled, <cr> will cancel completion mode preserving the current
> text.

https://github.com/ervandew/supertab/blob/master/doc/supertab.txt