Vim: rename some plugins
I decided that I want the directory names of my Vim plugins to match the
names of their repositories. I think this will make it easier for me to
understand exactly which plugins I am using. In this commit I rename a
few of them that I found that were not named after their repositories.