Vim: change final placeholders for Jasmine snippets
I found that when I had a `$2` here my tabbing was stuck in the snippet
tabstops longer than I expected.

The `$0` is a special tabstop that is always the last tabstop in the
snippet no matter how many tabstops are defined. Using this instead
seems to improve when UltiSnips knows when I am done with tabstops and
goes back to normal tabbing behavior.