Vim: tell vim-jsx that the JSX pragma is not required
A recent commit [1] in vim-jsx made it so that the JSX syntax
highlighting is only enabled if the JSX pragma is found in the file.
Since we use a webpack loader that automatically prepends the pragma to
our JSX files, this broke syntax highlighting for me.

Thankfully, an option was also added to disable this check [2], which I
have added to my Vim configuration.

In the issue discussion [3], it looks like this might break HTML syntax
highlighting, but I so rarely edit a regular HTML document in Vim that
I'm not worried about this problem.

[1]: https://github.com/mxw/vim-jsx/commit/d04fd5ec463c92815f89fbe3aae29b92272e3d8a
[2]: https://github.com/mxw/vim-jsx/commit/1e50d3fd1d4a061641a059fe9d6589b143758472
[3]: https://github.com/mxw/vim-jsx/issues/9