Slate: update internal resolution for 13" Retina Macbook Pro
When I switched to this laptop, my autoswitching Slate configuration
stopped working because my screen has a different resolution. After some
Googling, it seems that the logical resolution of this display is
1280x800. I decided to update my configuration for "internal" instead of
adding a new alias was the way to go because I don't plan on going back
to my old laptop.