Vim: add Henric's import-js plugin
@trotzig recently cooked up a Vim plugin that will hopefully save me
some time and effort when writing JavaScript. To use it, place your
cursor over a variable name and activate the plugin. By default it is
bound to <Leader>-j. This will search through your project for the
correct JavaScript file and add it to the set of required modules at the
top of the file you are working in.