Vim: add vim-dispatch
I have configured this plugin to run RSpec specs for the currently
opened RSpec file when running `:Dispatch` and to run all Jasmine specs
when in a Jasmine file.

I would like to have Jasmine scoped to the currently opened file, but
that is currently not a feature that the Jasmine rake task supports. I
have submitted a feature request to the jasmine-gem project to add
filename scoping:

  https://github.com/jasmine/jasmine-gem/issues/234

In the meantime, I will have to settle for scoping manually by using the
`fdescribe` and `fit` features in Jasmine 2.1.0.

Additionally, there is currently a problem affecting the combination of
Spring and Rake versions that I am using that breaks Rake with Spring.
To get around this, I am not using Spring for my Jasmine runs at this
time. More info:

  https://github.com/rails/spring/issues/366