Zsh: Set USER to `whoami` if it has not been set
When running, I was seeing an error message like:

> mkdir: `/tmp/`: operation not permitted

After some digging, this appears to be happening inside of tmux because
Dispatch creates new panes and windows with `env -i`, which completely
empties the environment. Since the TMPDIR variable setting I get from
Prezto uses $USER, and since $USER was not being set correctly, this was
no longer working as expected.

To fix this, I just need to set the USER variable to `whoami` when zsh
starts up, if it hasn't been set yet.

More info:

- https://github.com/tpope/vim-dispatch/issues/64
- https://github.com/sorin-ionescu/prezto/pull/605