Vim: rename pascalCase snippet helper to pascal_case_basename
I am about to add another function here and realized that the name I had
originally given to this one wasn't as good as I like it to be. Also, I
realized that the naming convention in Python is snake_case, not
camelCase, so I made that adjustment at the same time.