Git: remove whitespace fix
This setting makes checking in third-party code, such as that added by
npm or bower, a chore because it strips trailing whitespace when
committing but leaves it in the working tree. Since I am using
overcommit to guard against whitespace problems where I care about them,
I don't really get much value out of this option and am removing it.