Vim: disable syntastic when running :Acks
I have :Acks set up to make multi-file find/replace a little more
comfortable in Vim. However, it was running slowly because syntastic was
running after saving each file. Since I am not actually opening the
files and looking at them, I wasn't getting any value out of this and
want to just disable it.

I originally looked for a command that explicitly disables it instead of
toggling it, since the original state could be off (which would turn it
on for the :Acks run), but that doesn't seem to exist so I am just going
to run with :SyntasticToggleMode for now and see if I run into any
issues. Since I am not usually turning it off, I don't expect to run
into any.