Vim: add some JSX propTypes snippets
I've been using these to make me a little more productive when writing
React components.

I got pretty fancy with one by adding some autocompletion to it, and
I've been on the fence about it, but I figured I'd put it in here for
now and see how it goes.

While I was at it, I fixed a number of warnings that flake8 told me
about in my Python code.