Vim: add `desf` Jasmine snippet for top-level `describe()`
`this` is `undefined` at the top level in modules. This is problematic
when using ES6 arrow functions at the top level, as I am wont to do in
my Jasmine specs. This commit adds a snippet that allows me to create
a non-arrow function version of my `describe` block for the top level of
my Jasmine specs.