Vim: improve file <Tab> auto-completion
The `wildmenu` setting tells Vim to bring up a menu of the options when
trying to auto-complete a filename (such as when pressing <Tab> while
running `:e`). This will help me understand what is happening what I try
to auto-complete a filename.

The `wildmode` option with these settings instructs Vim to operate more
similarly to the auto-complete experience I have in my shell: fill in
characters up to the longest unambiguous portion.

These two settings combined make file auto-completion in Vim much more
pleasant and easier to work with.