Vim: move JSX ftdetect into `after`
The vim-jsx plugin contains an ftdetect file that sets the filetype on
*.jsx files to `javascript`. Unfortunately, this ends up clobbering my
local ftdetect file that sets the filetype to `javascript.jsx` or
`javascript.jasmine.jsx` which I rely on for better scoping of my
settings.

Moving this into the `after` directory allows my configuration to take
precedence.

Some more context on this issue can be found at:

  https://github.com/mxw/vim-jsx/issues/29
  https://github.com/mxw/vim-jsx/pull/30