Tmux: start in ~ when creating windows
I had previously added this line to maintain the current directory when
creating new windows and panes in tmux; however, I have found it to be
more intuitive to maintain the current directory when creating new panes
but not new windows. I'm going to try removing this so tmux will have
the default behavior of landing me in my home directory when creating
new windows and see how it goes.