Tmux: set up Tmuxinator
Tmuxinator is a gem that allows me to have a configuration file to start
up a Tmux session with a specific layout. Since I always do the same
layout manually, this should save me a good amount of time going
forward.

When setting this up, I first ran into an issue where it wasn't running
the commands in the panes that I told it to. After exiting my session, I
noticed error messages like:

> can't find pane: 1

It turns out that since I am setting the base-index for windows to 1,
Tmuxinator also requires that I set the base-index for panes to 1, so I
did that.

More info:

  - https://github.com/tmuxinator/tmuxinator/issues/137
  - https://github.com/tmuxinator/tmuxinator/issues/225

Once I solved that, I was off to the races and just needed to configure
my session the way I like it. In this process, I refined a few things
that I normally do, in order to make the automated version a bit
smoother.

- I switched from `mysqld` to `mysql.server start` because it will work
  more smoothly if the server is already started. Similarly with
  including the `--daemonize yes` option on `redis-server`.

- I also found a more graceful way to kill the runaway jasmine server
  that we've been struggling with by killing whatever is running on port
  8888 instead of what I was previously using (either `pkill -f jasmine`
  or more recently `pkill -f rake`).

- I decided to try out running `:CommandT` when Vim starts, because this
  is what I want 99% of the time, and it will give Command-T an
  opportunity to warm up its index.

To install Tmuxinator, run:

  gem install tmuxinator

To start my brigade session, run:

  tmuxinator start brigade

or simply:

  mux brigade