Tmux: fix hot modules configuration
It turns out that we need these environment variables in all of these
processes, so I'm adding it to all of these commands.

I wonder if there is a nicer way to do this that can still be local to
these processes, but I don't want to take the time right now to figure
that out.