Dotfiles: use `ln -sFi` in init script
Adding the -Fi flags will prompt to overwrite existing files and
directories when attempting to create symlinks instead of simply
failing.

According to the manpage for `ln`, `-i` is non-standard and use in
scripts is not recommended. Also `-F` is FreeBSD extension and should
not be used in portable scripts. However, they seem to work just fine on
my Mac right now and that's the environment where I will most likely be
running this script, so I'm going to live a little on the wild side and
see what happens.