Shell/Vim: switch from Solarized to Base16 colors
I've been considering switching from Solarized to a different
colorscheme for a while now. I recently learned of Base16 and liked the
look of it its configurability. It comes with different sets of colors
that you can easily switch between, including one that is similar to
Solarized.

Most of this configuration was taken from @wincent. It allows the colors
to be controlled in one place for both the shell and Vim, which is
great. To switch to a different set of colors, there are two commands,
`light` and `dark` that take an argument of the name of colors. Example:

  dark monokai

To see a list all schemes, run `dark help`.

More information can be found at:

- https://github.com/wincent/wincent/commit/35f61972a441e7cf2a696393184ff9063c95c5cd

The colors aren't quite right in iTerm 2 yet (compared to MacVim and the
demo page, the background isn't as dark as it should be, and maybe some
other differences), but I haven't found a solution yet.