Vim: prefer local eslint for Syntastic
I find Syntastic to be a valuable tool, but there isn't a built in way
to set it up to prefer locally installed npm packages for checker
executables. This means that I need to keep my global version at the
same place as my local version. Although this is annoying, it can
actually end up being problematic if I am switching between projects
that use different versions.

To make Syntastic prefer the local version, I am adding an executable
that is in my PATH that adds the npm bin to the bath while trying to
execute eslint.

I wanted to do this in a more general way similar to what @lcd047
suggested in https://github.com/scrooloose/syntastic/issues/1247

- add an `npm-exec` alias or script in my PATH that just does
  `PATH=$(npm bin):$PATH "$@"`
- set the syntastic executable to `npm-exec eslint`

Unfortunately Syntastic doesn't like either the alias or the script with
the argument, so I ended up just adding an eslint-specific script. This
is a bit of a bummer, but I don't expect to have many of these anyway,
so this seems like an acceptable tradeoff.