sh: Prevent italics installation failure if tmux terminfo doesn't exist
I am setting up a new laptop and had my extras install script was
failing with:

> infocmp: couldn't open terminfo file (null).

I realized that this was happening because the tmux terminfo file did
not exist on this system. Since this first line is just reading it in as
a backup, we can just keep moving forward without any problems.