Vim: map <Leader>d to :Dispatch
I am setting up vim-dispatch in a new project and decided that it would
be nice to have a mapping to easily invoke the :Dispatch command. I am
also trying out mapping <Leader>q to :q to make it easier to close the
window it creates, but I'm a little hesitant to commit it just yet.