Scripts: chown ~/.viminfo in init
In setting up a new machine, I noticed that Vim history was not saving
across sessions. I was able to track this down to not being able to
write to ~/.viminfo. I hope that adding this to my init script will
prevent this from happening next time.

I included touching the file first in case it was not there, and I
included the `|| true` to prevent it from erroring the init script out
if it already exists and I don't have permissions to touch.