Vim: make <Ctrl-^> go to last file, not to netrw
It has long annoyed me that once I've started to explore the directory
of a file using vim-vinegar's `-` mapping, I can't easily get back to
the file with `<space><space>` (ie. `<leader><leader>`) or `<C-^>`.
Thanks to @wincent [0], I have found a workaround:

  https://github.com/tpope/vim-vinegar/issues/25

Unfortunately, this doesn't fix my other annoyance of netrw polluting
the jumplist. Looking at `:help netrw` it looks like `g:netrw_keepj`
intends to do this and it is defaulted to "keepj" (on), but it isn't
quite as good as I would like.

[0]: https://github.com/wincent/wincent/commit/337792ff