Vim: tweak some Jasmine and RSpec snippets
I recently added a JSX `con` snippet for the constructor, so I needed to
disambiguate the `context` snippets. To keep things consistent between
RSpec and Jasmine, I decided to change them both to `cont`. This is
awfully close to `const` which I type all the time, so we'll see how it
goes.

The other change I am making is to stop using `this` for the subject
pattern. Since we are using modules everywhere, this can be simplified
some.