zsh: Avoid loading prezto for non-interactive shells
While working on an Atom plugin that shells out to get environment
information, I noticed that that operation was pretty slow for me. I'm
using prezto, which adds some nice features to my sells, but I only need
them in interactive shells. It turns out that there is a way to detect
if the shell is interactive, which I can use to prevent loading prezto.
On my laptop, this seems to shave off about 300ms, which I think is a
pretty nice win!

http://unix.stackexchange.com/questions/26676/how-to-check-if-a-shell-is-login-interactive-batch/26782#26782