Vim: map <C-direction> to <Esc>direction
I have overloaded my capslock key to be Esc when pressed by itself and
Ctrl when chorded. I believe that there is a short window of time after
I have released the key where it still considers it chorded, which
occasionally causes me some issues when I am typing very fast.

One of these issues happens when I am in insert mode and I want to go
back to normal mode and start heading off in a direction by pressing `j`
or `k`. Since this is sometimes interpreted as <C-j> or <C-k> instead of
<Esc>j or <Esc>k, I end up triggering a different feature than I
intended.

This happened to be compounded by the way I had configured YouCompleteMe
to allow me to use <C-j> and <C-k> to control the popup window.

YouCompleteMe will map the keys you give it to control the popup window
if it is open, or to do the default if it is not open.

  https://github.com/Valloric/YouCompleteMe/blob/f44435b8/autoload/youcompleteme.vim#L219-L220

I'm not entirely sure why at this time, but this seems to clobber any
mappings that I already had on those keys--even if I used `imap` instead
of `inoremap`. So, I decided to work around this by removing <C-j> and
<C-k> from this list and handling the pumvisible expression myself with
my special fallback. I don't really like this because of the
duplication and spreading out of my configuration that does essentially
the same things, but it gives me the functionality that I want.

I am considering moving all of these mappings out of my autocomplete
configuration and into my mappings file, but I figured I'd roll with
this for a bit and see how it goes.