Vim: embolden filename in statusline
Shamelessly stolen from @wincent's dotfiles. This was nice and easy,
thanks to his pinnacle plugin.

The only real bummer is that in my current typeface (Inconsolata), the
difference between bold and medium is quite subtle. Perhaps I need to
find a new typeface with a regular weight.

I'm holding out hope for ligature support in iTerm [0] and I've heard
good things about Fira Code [1] but it does not have italics [2], so that's
a non-starter at least until Mozilla adds italics to Fira Mono.

Monoid [4] and Hasklig [5] have italics and ligatures and look pretty
promising. I've been using Consolas/Inconsolata since 2006 so it feels
like a pretty big change, but perhaps it is a valuable one.

[0]: https://gitlab.com/gnachman/iterm2/issues/3568
[1]: https://github.com/tonsky/FiraCode
[2]: https://github.com/tonsky/FiraCode/issues/172
[3]: https://github.com/mozilla/Fira/issues/38
[4]: http://larsenwork.com/monoid/
[5]: https://github.com/i-tu/Hasklig