Vim: add number of errors and line of first error to statusline
These come from syntastic. I tried to find a way to use the powerline
glyph, but I couldn't see an easy way to make the glyph work with
highlight groups in a way that it would only appear when there were
errors, so I decided to just go with something simple for now.

This will look something like

  1 ✖ :6

which means that there is 1 error and it is on line 6. I think this will
help me work more efficiently.