Update Base16 submodules, add background detection
A recent update [0] to the Base16 repos removed the nice foo.dark.sh and
foo.light.sh naming scheme that we were relying on in these scripts to
help ensure that the background in Vim was always correctly set to
either dark or light.

[0]: https://github.com/chriskempson/base16-shell/commit/eeeff9de8

Although some of the light themes have been brought back [1], they do
not maintain this naming scheme.

[1]: https://github.com/chriskempson/base16/issues/42

So I decided to extract the background color from the theme and compute
its luma to determine how perceptually dark it is. 0 is black and 255 is
white, so it seems that anything less than 127.5 should be considered
"dark" and anything above that should be considered "light".

This also means that the nice `dark` and `light` helper functions aren't
very useful to us anymore, so I switched things around to just use
`color`.

I'm not the best at shell scripting, so there may be some room for
improvement here, but this seems to get the job done and I find it to be
pretty readable.

Worth noting is the scripts in the base16-shell repo are now no longer
executable, so I needed to change my script to use `source`. This seems
to have the side-effect of ringing the bell and I was unable to figure
out how to prevent this from happening. There is probably an easy way,
but it is unknown to me and difficult to Google for. I hope to come back
and silence this at some point.