Vim: improve listchars
It looks like @wincent found some nicer characters to use here. The nbsp
and tab in particular seem to be improved. I figured I take them for a
spin.

At the same time, I stole the comments describing what these characters
are. I'm not sure if they will ever be useful, but maybe?