Git: add git-who script
I sometimes just want a list of the people who have touched a file, so
I've written a quick-and-dirty shell script that will tell me this. I
have set up my init script to symlink it into ~/bin, which is in my
PATH, so git will pick it up as a sub-command. It can be run like:

  git who -- path/to/file

I wanted to sort it by most recent authors, but I don't know of an easy
way to do that at this time, so it is currently sorted alphabetically.
If I want to sort it by most commits, I can pipe it to sort.

  git who -- path/to/file | sort