Vim: add current buffer's file size to statusline
I often find that I want to know the size of the file that I am
currently editing. I figured it would be nice to surface this in the
statusline. The only bummer is that this number doesn't update until I
save, but I think I can live with that.