Vim: improve format of wrapped long lines
These options will preserve the previous level of indentation on wrapped
lines, while indenting them an extra couple of spaces for emphasis. I'm
not entirely sure I need both the extra indentation and the showbreak
character here (or either of them, really, but I'm going to roll with
this for a bit and see how it feels.

Stolen from:

- https://github.com/wincent/wincent/commit/d8d586c3
- https://github.com/wincent/wincent/commit/5311ed53