Vim: strip trailing semicolons in console.log snippet
I often use my console.log snippet by using visual block mode to select
a line of JavaScript to wrap it in a console.log statement. I always end
up with a trailing semicolon in my console.log statement this way, which
I have to manually remove. I'm tired of doing this, so I am automating
it by using a little Python to remove the trailing semicolon from the
line when using this snippet.