Git: enable precomposeunicode
I am working in a codebase that has some filenames with unicode
characters. Without this setting, I end up with some weird situations,
such as the file being untracked and then when I delete it, it being
suddenly tracked. Enabling precomposeunicode allows this to work better
for me. From the docs:

> This option is only used by Mac OS implementation of Git. When
> core.precomposeUnicode=true, Git reverts the unicode decomposition of
> filenames done by Mac OS. This is useful when sharing a repository
> between Mac OS and Linux or Windows. (Git for Windows 1.7.10 or higher
> is needed, or Git under cygwin 1.7). When false, file names are
> handled fully transparent by Git, which is backward compatible with
> older versions of Git.

https://git-scm.com/docs/git-config