Vim: Add ale plugin
I've decided to try out this plugin as a replacement for Syntastic.
Importantly, it works asynchronously, which should give me faster
feedback from my linters.

This commit brings in some basic configuration that I think is a pretty
good start. I wanted to drop in line the number of the error on the
statusline, but I didn't see an easy way to do that right now so I
decided to punt on it. Perhaps I'll add some mappings to the provided
navigation commands, e.g.:

  nmap <silent> <C-k> <Plug>(ale_previous_wrap)
  nmap <silent> <C-j> <Plug>(ale_next_wrap)

More documentation on this can be found at `:help
ale-navigation-commands`.