Ruby: Switch from rbenv to chruby
As suggested by @wincent in 2cbf354#commitcomment-21001197, I'm going to
give chruby a go. He's been using it since
https://github.com/wincent/wincent/commit/d5d5736a and I trust his
advice. It appears to be simpler, which is nice--particularly because I
try to use Ruby as little as possible these days.

Likewise, ruby-install seems to be the way to install Ruby for chruby so
I swapped out ruby-build for that.

Fixes #5