OSX: Update key repeat rate
I was forced to update to Sierra at work, which broke Karabiner, which I
used for important things such as adjusting my key repeat rate. After
some digging, I discovered that I can use these commands to set it close
to what I had it at.

Unfortunately it seems that setting it to `-float 1.8` rounds it down to
`-int 1`, which is too fast for me. And `-int 2` is too slow, but better
than 1 so that's what I'm rolling with.