Hammerspoon: add magic mappings for capslock
This was basically stolen from @wincent.

https://github.com/wincent/wincent/blob/16b86434/roles/dotfiles/files/.hammerspoon/eventtap.lua

The only real modification I made in eventtap.lua is changing line 256
from

  tapped = 'delete'

to

  tapped = 'escape'

I've grown pretty accustomed to having escape on the home row,
especially since I use the Microsoft ergo keyboard with the worst escape
key ever. Also, if the trend of removing the escape key from the
keyboard continues, this will be even more critical.

I didn't take the time to fully understand this code, unfortunately, but
it appears to also add a magic ctrl mapping for return so I have some
symmetry. I doubt I'll really make much use of this, but I figured I'd
leave it in there since it was already done.