Vim: auto-change '->' to '→' in git commit messages
I type this arrow a fair amount when writing commit messages for
updating dependencies, and I often run out of characters. I can save
some characters by using the unicode arrow instead, and this
abbreviation will help me automatically use it.

This might turn out to be a terrible idea, but we'll see how it feels
for a while.