Add vscode settings
I am taking vscode for a spin, and I want to keep my settings in my
dotfiles. This is pretty rough for now, but it seems to be a good start.

This directory needs to be symlinked to

  ~/Library/Application Support/Code/User

I'll probably want to wire this up in my init script, but for now,
I'll just go with a manual symlink I created myself.