zsh: avoid chained git aliases
Instead of using a zsh alias to point to a git alias, I am removing a
link in the chain by going directly to the final command here.

I've been experiencing some friction with git lately having index.lock
files when I try to commit and it seems to happen much more frequently
when I use these zsh aliases, so I was hoping this would help resolve
that a little. Unfortunately, it does not seem to have helped, but this
still seems like a positive change.