vscode: Configure breadcrumbs to work like vim-vinegar
I've been using Vim for a while now and I'm trying out Visual Studio
Code. One thing I have been missing a lot is the ability to navigate
between files in the way that vim-vinegar enables.

Using this plugin in Vim, I can press - and be taken into a file
browser, starting in the directory that the current file is in.
Selecting a file opens it, and pressing - again traverses up a directory
level.

https://stackoverflow.com/questions/54834068/replacement-for-vim-vinegar-in-vs-code

I think I can replicate the experience I want using breadcrumbs. This
commit represents my first attempt at this by:

- enabling breadcrumbs
- adding a vim normal mode keybinding of `-` to enable and focus
  breadcrumbs
- adding a keyboard shortcut of `-` while breadcrumbs are visible to
  navigate "up a level"

This isn't perfect yet, specifically pressing enter on a directory does
nothing, but it is a big step in the right direction.