Make submodule updating script more resilient
I was trying to update submodules today and ran into some problems
because git was complaining like:

```
There is no tracking information for the current branch.
Please specify which branch you want to merge with.
See git-pull(1) for details.

    git pull <remote> <branch>

If you wish to set tracking information for this branch you can do so with:

    git branch --set-upstream-to=origin/<branch> master
```

However these submodules were tracking a remote tag, and not a branch,
so there's really no good way to resolve this that I know of.

Adding `|| :` to the script allows it to continue on to the next
submodule if there are any errors.