vscode: disable accessibility support
This option defaults to auto, which makes vscode detect if a screen
reader is in use and then adjusts the interface in a number of ways.

I sometimes use a screen reader to do accessibility testing for web
development work, but I don't actually want my editor to adjust when
this is on.

https://code.visualstudio.com/docs/editor/accessibility#_screen-reader-mode