git: add delete-merged alias
I was relying on my shell history to do this, but since I do it often
enough I figured that I should probably have an alias for it. This makes
it more portable for me as well.

While I was doing this, I decided to incorporate the same trick I used
for my refresh alias to automatically determine origin's default branch,
which will make this even more portable as I switch between repos.