vscode: Improve compatibility with copilot and markdown
I noticed that tab-completion for GitHub Copilot was not working for me
in markdown files. I found a workaround that someone posted and it seems
to do the trick for me.

https://github.com/yzhang-gh/vscode-markdown/issues/1011

In my reading of the linked issue it seems that a fix may eventually
land in the extension in which case I should be able to remove this
keybinding.