zsh: Activate mise in .zshrc
I am looking for a node version manager as I set up my new laptop. I
have been been using nvm for ages, but it is feeling a little long in the tooth.

asdf is a version manager for all languages, which is appealing. mise
also looks good, which seems to be a rewrite of asdf but in Rust for
performance and also improving on the API.

https://mise.jdx.dev/dev-tools/comparison-to-asdf.html

I'm going to give mise a shot and see how it feels. I installed it by
running:

```
curl https://mise.run | sh
```

Getting started guide:

https://mise.jdx.dev/getting-started.html